% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_pred_dist.R
\name{alpha_pred_dist}
\alias{alpha_pred_dist}
\title{Predicted Distance Function(Alpha-only)}
\usage{
alpha_pred_dist(data, startyear, endyear)
}
\arguments{
\item{data}{'data' should be a data object with following variables: YEAR (the year of event), EVENT_DATE (the date of event),
ACTOR1 (focal actor name), ACTOR2 (alter actor name), LATITUDE (latitude), LONGITUDE (logitude). Note: the variable names should be EXACT.}

\item{startyear}{The first year you want to produce the predicted distance. It should be later than the earliest year of your data.}

\item{endyear}{The last year you want to produce the predicted distance. It should be earlier than the lastest year of your data.}
}
\description{
This function allows you to produce the predicted distance (alpha-only) of moving actors in your dataset
}
\examples{
data(acled)
data<-acled
pred_alpha<-alpha_pred_dist(data,1999,2000)
}
\keyword{practice}
