% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wt_vid.R
\name{wt_vid}
\alias{wt_vid}
\title{A What video game function}
\usage{
wt_vid(ask = TRUE)
}
\arguments{
\item{ask}{Do you want to ask him? Defaults to TRUE.}
}
\description{
This function allows you to ask Sangyeon what video game he plays nowadays.
}
\examples{
wt_vid()
}
\keyword{practice}
